refactor(settings): extract the general preferences page
Move browsing, refresh, video preview, and click-activation controls into
their own module so the settings layer only composes that page.
