+++ src/ui/settings/tests.rs
@@ -12,13 +12,14 @@ use crate::services::{
 use super::{
     CHANNEL_ORDER, COMPACT_NAVIGATION_BREAKPOINT, DIALOG_HEIGHT, DIALOG_MARGIN, DIALOG_WIDTH,
     RELEASE_CHANNEL_DESCRIPTION, RELEASE_CHANNEL_TITLE, UPDATE_DUE_INTERVAL, aur_update_command,
-    channel_index, effective_update_channel, force_due_update_check, install_guard,
-    installed_version_status, is_stale_check, managed_channel_description, managed_install_summary,
-    offer_still_eligible, omarchy_update_command, resolve_update_method_async,
-    responsive_dialog_size, shows_available_release_notes,
+    channel_index, effective_update_channel, force_due_update_check,
+    general::{video_preview_backend_label, video_preview_control_state},
+    install_guard, installed_version_status, is_stale_check, managed_channel_description,
+    managed_install_summary, offer_still_eligible, omarchy_update_command,
+    resolve_update_method_async, responsive_dialog_size, shows_available_release_notes,
     theme::{theme_background_is_light, theme_name_matches},
     update_check_due, update_check_message, update_dialog_status, update_status_markup,
-    uses_compact_navigation, video_preview_backend_label, video_preview_control_state,
+    uses_compact_navigation,
 };
 use crate::sandbox::MediaPreviewBackend;
